Moved query parameter exporter to Ghost core
no ref

Keeping the exporter alongside the canonical policy lets it use Ghost's existing TypeScript runtime and avoids adding tooling dependencies at the workspace root.
